+++ src/components/SavedSearchItem.tsx
@@ -13,10 +13,10 @@ import {
 	useToast,
 	Text,
 	Stack,
-	Wrap,
-	Spacer,
 	StackItem,
 	Flex,
+	IconButton,
+	ButtonGroup,
 } from '@chakra-ui/react';
 import { isEqual } from 'lodash';
 import { FiSearch, FiSave, FiDelete, FiEdit2 } from 'react-icons/fi';
@@ -27,7 +27,7 @@ import useCandidateSearch from '@hooks/queries/useCandidateSearch';
 import useTaxonomyTerms from '@hooks/queries/useTaxonomyTerms';
 import useViewer from '@hooks/queries/useViewer';
 import useSaveSearch from '@hooks/mutations/useSaveSearch';
-import useDeleteSavedSearch from '@hooks/mutations/useDeleteSavedSearch';
+import useDeleteOwnSavedSearch from '@hooks/mutations/useDeleteOwnSavedSearch';
 import TextInput from '@common/inputs/TextInput';
 import SearchParamTags from '@common/SearchParamTags';
 import ConfirmActionDialog from '@common/ConfirmActionDialog';
@@ -49,16 +49,14 @@ export default function SavedSearchItem({ id, title, searchTerms, ...props }: Pr
 		searchDispatch,
 	} = useContext(SearchContext);
 
-	// const { openDrawer } = useContext(SearchDrawerContext);
-
-	const isNamed = title && title !== '';
+	const hasName = title && title !== '';
 
 	const { isOpen: editIsOpen, onOpen: editOnOpen, onClose: editOnClose } = useDisclosure();
 	const { isOpen: deleteIsOpen, onOpen: deleteOnOpen, onClose: deleteOnClose } = useDisclosure();
 	const initialSaveModalRef = useRef(null);
 
 	const { saveSearchMutation } = useSaveSearch();
-	const { deleteSavedSearchMutation } = useDeleteSavedSearch();
+	const { deleteOwnSavedSearchMutation } = useDeleteOwnSavedSearch();
 
 	const toast = useToast();
 
@@ -87,15 +85,18 @@ export default function SavedSearchItem({ id, title, searchTerms, ...props }: Pr
 				filterSet,
 			},
 		});
-
-		// openDrawer();
 	};
 
 	const handleEditClick = (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
 		editOnOpen();
 	};
 
-	const handleRenameSubmit = (e: FormEvent) => {
+	const handleEditOnClose = () => {
+		setSaveSearchFieldText(title ? title : '');
+		editOnClose();
+	};
+
+	const handleSave = (e: FormEvent) => {
 		e.preventDefault();
 
 		saveSearchMutation({
@@ -115,7 +116,6 @@ export default function SavedSearchItem({ id, title, searchTerms, ...props }: Pr
 			});
 
 			setSaveSearchFieldText('');
-
 			editOnClose();
 		});
 	};
@@ -127,7 +127,7 @@ export default function SavedSearchItem({ id, title, searchTerms, ...props }: Pr
 	const handleDelete = () => {
 		if (!id) return;
 
-		deleteSavedSearchMutation(id.toString(), loggedInId).then(() => {
+		deleteOwnSavedSearchMutation(id.toString(), loggedInId).then(() => {
 			deleteOnClose();
 
 			toast({
@@ -144,57 +144,66 @@ export default function SavedSearchItem({ id, title, searchTerms, ...props }: Pr
 	return termIds && termIds.length > 0 ? (
 		<>
 			<Stack w='auto' alignItems='space-between' {...props}>
-				<StackItem display='flex' alignItems='center' gap={1}>
-					<Text my={0} fontSize='lg'>
-						{isNamed ? (
+				<StackItem>
+					<Flex>
+						{hasName ? (
 							<LinkWithIcon
 								onClick={handleEditClick}
 								icon={FiEdit2}
-								iconSide='right'
-								iconProps={{ boxSize: 3, mb: '2px', ml: 1 }}
+								fontSize='lg'
+								my={0}
+								mr={8}
+								flex={1}
+								iconSide='left'
+								color='inherit'
+								borderBottomWidth='2px'
+								borderBottomStyle='dotted'
+								textDecoration='none !important'
+								_hover={{ borderBottom: '1px  dotted brand.blue' }}
+								_light={{
+									borderBottomColor: 'gray.300',
+									_hover: { borderBottomColor: 'gray.400' },
+								}}
+								_dark={{ borderBottomColor: 'gray.600', _hover: { borderBottomColor: 'gray.400' } }}
+								iconProps={{ boxSize: 4, mb: '2px', ml: 1, position: 'relative', top: '2px' }}
 							>
 								{title}
 							</LinkWithIcon>
 						) : (
 							false
 						)}
-					</Text>
-					<Spacer />
-					<StackItem as={Wrap} alignItems='center' spacing={1}>
 						{id ? (
-							<Button
-								leftIcon={<FiSearch />}
-								colorScheme='green'
-								aria-label='Rerun this search'
-								title='Rerun'
+							<ButtonGroup
+								alignItems='center'
+								justifyContent='space-between'
+								flex='0 0 auto'
 								size='sm'
-								onClick={handleSearchClick}
+								spacing={1}
 							>
-								Search
-							</Button>
+								<IconButton
+									icon={<FiSearch />}
+									colorScheme='green'
+									aria-label='Reuse this search'
+									title='Reuse this search'
+									onClick={handleSearchClick}
+								/>
+								<IconButton
+									icon={<FiDelete />}
+									colorScheme='orange'
+									aria-label='Delete this search'
+									title='Delete'
+									onClick={deleteOnOpen}
+								/>
+							</ButtonGroup>
 						) : (
 							false
 						)}
-						{id ? (
-							<Button
-								leftIcon={<FiDelete />}
-								colorScheme='orange'
-								aria-label='Delete this search'
-								title='Delete'
-								size='sm'
-								onClick={deleteOnOpen}
-							>
-								Delete
-							</Button>
-						) : (
-							false
-						)}
-					</StackItem>
+					</Flex>
 				</StackItem>
 				<StackItem>
 					<Flex w='full' justifyContent='space-between'>
 						<SearchParamTags termIds={termIds} termItems={terms} />
-						{!isNamed ? (
+						{!hasName ? (
 							<Button
 								colorScheme='blue'
 								leftIcon={<FiSave />}
@@ -213,18 +222,22 @@ export default function SavedSearchItem({ id, title, searchTerms, ...props }: Pr
 				</StackItem>
 			</Stack>
 
-			<Modal initialFocusRef={initialSaveModalRef} isOpen={editIsOpen} onClose={editOnClose}>
+			<Modal initialFocusRef={initialSaveModalRef} isOpen={editIsOpen} onClose={handleEditOnClose}>
 				<ModalOverlay />
 				<ModalContent>
-					<ModalHeader>{isNamed ? 'Rename this search' : 'Save this search'}</ModalHeader>
+					<ModalHeader pb={2}>{hasName ? 'Rename this search' : 'Save this search'}</ModalHeader>
 					<ModalCloseButton />
 
 					<ModalBody pb={6}>
-						<Text>Give this search a descriptive name to easily re-run this search.</Text>
+						<Text fontSize='sm' mt={0}>
+							Give this search a short, descriptive name to easily run it again.
+						</Text>
 
-						<form id='rename-search' onSubmit={handleRenameSubmit}>
+						<form id='rename-search' onSubmit={handleSave}>
 							<FormControl>
-								<FormLabel>Name</FormLabel>
+								<FormLabel aria-label='Name' visibility='hidden' position='absolute' left='9000px'>
+									Name
+								</FormLabel>
 								<TextInput
 									name='title'
 									placeholder='My search'
@@ -233,10 +246,17 @@ export default function SavedSearchItem({ id, title, searchTerms, ...props }: Pr
 									ref={initialSaveModalRef}
 								/>
 							</FormControl>
-							<Button colorScheme='blue' mr={3} type='submit'>
+							<Button
+								colorScheme='blue'
+								mr={3}
+								type='submit'
+								isDisabled={saveSearchFieldText === title}
+							>
 								Save
 							</Button>
-							<Button onClick={editOnClose}>Cancel</Button>
+							<Button onClick={handleEditOnClose} colorScheme='red'>
+								Cancel
+							</Button>
 						</form>
 					</ModalBody>
 				</ModalContent>