+++ src/components/SavedSearchItemList.tsx
@@ -12,11 +12,13 @@ export default function SavedSearchItemList({ ...props }: Props) {
 	const [savedSearches] = useSavedSearches();
 	const MotionBox = motion(chakra.div);
 
+	// TODO Maintain indexes when editing a saved search's name.
+
 	return (
 		<chakra.div {...props}>
 			{savedSearches && savedSearches.length > 0 ? (
 				<Container ml={0} px={0} maxW='4xl'>
-					<Flex gap={4} justifyContent='space-between' position='relative'>
+					<Flex gap={4} justifyContent='space-between' position='relative' flexWrap='wrap'>
 						<AnimatePresence>
 							{savedSearches.map((savedSearch: WPPost) => {
 								const { id, content, title } = savedSearch;
@@ -39,9 +41,8 @@ export default function SavedSearchItemList({ ...props }: Props) {
 											my={0}
 											py={2}
 											px={3}
+											variant='gray'
 											w='auto'
-											_dark={{ bgColor: 'gray.800' }}
-											_light={{ bgColor: 'gray.100' }}
 										>
 											<SavedSearchItem searchTerms={filters} title={title} id={id} />
 										</Card>