+++ src/theme/components/Text.ts
@@ -19,9 +19,9 @@ const variants = {
 	}),
 	helperText: defineStyle({
 		lineHeight: 'normal',
-		fontSize: '2xs',
+		fontSize: 'xs',
 		letterSpacing: '0.8px',
-		opacity: 0.74,
+		opacity: 0.8,
 		_dark: {
 			color: 'text.light',
 		},