+++ src/views/SearchResultsView.tsx
@@ -1,10 +1,11 @@
 import { useContext, useMemo } from 'react';
-import { Text } from '@chakra-ui/react';
-import { flattenfilterSetPositions } from '@lib/utils';
+import { Flex, IconButton, Text } from '@chakra-ui/react';
+import { prepareSearchFilterSetForSave } from '@lib/utils';
 import { SearchContext } from '@context/SearchContext';
 import TextCenterline from '@common/TextCenterline';
 import SavedSearchItem from '@components/SavedSearchItem';
 import CandidateList from '@components/CandidateList';
+import { FiCalendar } from 'react-icons/fi';
 
 export default function SearchResultsView() {
 	const {
@@ -13,6 +14,8 @@ export default function SearchResultsView() {
 			results,
 		},
 	} = useContext(SearchContext);
+
+	const { jobDates } = filterSet;
 
 	const resultsCount = results.length;
 
@@ -38,11 +41,35 @@ export default function SearchResultsView() {
 		return sortedResults.map((item) => Number(item.id));
 	}, [results]);
 
+	const ConflictDateLegend = () => {
+		return jobDates && jobDates.startDate ? (
+			<Flex justifyContent='flex-start' alignItems='center' gap={1} mb={4} mt={0} ml={12}>
+				<IconButton
+					icon={<FiCalendar />}
+					variant='inline'
+					title='Search'
+					bgColor='red.300'
+					color='text.dark'
+					aria-label='Possible scheduling conflict'
+					size='xs'
+				/>
+				<Text variant='helperText' fontSize='sm'>
+					= Possible scheduling conflict
+				</Text>
+			</Flex>
+		) : (
+			<></>
+		);
+	};
+
 	return (
 		<>
-			<SavedSearchItem searchTerms={flattenfilterSetPositions(filterSet)} />
+			<SavedSearchItem searchTerms={prepareSearchFilterSetForSave(filterSet)} />
 			{resultsCount ? (
-				<TextCenterline fontSize='xl'>{resultsString()}</TextCenterline>
+				<>
+					<TextCenterline fontSize='xl'>{resultsString()}</TextCenterline>
+					{jobDates && jobDates.startDate ? <ConflictDateLegend /> : false}
+				</>
 			) : (
 				<Text fontSize='sm'>Your search results will appear here after you Search.</Text>
 			)}